% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridDonuts.R, R/tmapGridFlowers.R
\name{tmapGridDataPlot.tm_data_donuts}
\alias{tmapGridDataPlot.tm_data_donuts}
\alias{tmapLeafletDataPlot.tm_data_donuts}
\alias{tmapGridDataPlot.tm_data_flowers}
\alias{tmapLeafletDataPlot.tm_data_flowers}
\title{Internal methods for tm_flowers}
\usage{
\method{tmapGridDataPlot}{tm_data_donuts}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_donuts}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_flowers}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_flowers}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{shpTM, dt, gp, bbx, facet_row, facet_col, facet_page, id, pane, group, o, ...}{to be described}
}
\value{
internal tmap object
}
\description{
Internal methods for tm_flowers
}
\keyword{internal}
