% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layer_donuts.R
\name{opt_tm_donuts}
\alias{opt_tm_donuts}
\title{Options for tm_donuts}
\usage{
opt_tm_donuts(
  points.only = "ifany",
  icon.scale = 6,
  start = 0,
  direction = 1,
  inner = 0.4,
  just = NA,
  grob.dim = c(width = 48, height = 48, render.width = 256, render.height = 256)
)
}
\arguments{
\item{type}{cartogram type, one of: "cont" for contiguous cartogram, "ncont" for non-contiguous cartogram and "dorling" for Dorling cartograms}

\item{itermax, }{maximum number of iterations (see [cartogram::cartogram_cont()])}

\item{...}{arguments passed on to [cartogram::cartogram_cont()]}
}
\description{
Options for tm_donuts
}
