% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tmap.glyphs-package}
\alias{tmap.glyphs}
\alias{tmap.glyphs-package}
\title{Extension for tmap: glyphs can be created}
\description{
New layer functions available for 'tmap' that are used to create glyph maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-tmap/tmap.glyphs}
  \item \url{https://r-tmap.github.io/tmap.glyphs/}
  \item Report bugs at \url{https://github.com/r-tmap/tmap.glyphs/issues}
}

}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\concept{GIS}
\concept{bubble map}
\concept{choropleth}
\concept{statistical maps}
\concept{thematic maps}
