% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridDonuts.R
\name{tmapGridDataPlot.tm_data_donuts}
\alias{tmapGridDataPlot.tm_data_donuts}
\title{Internal methods for tm_donuts}
\usage{
\method{tmapGridDataPlot}{tm_data_donuts}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{shpTM, dt, gp, bbx, facet_row, facet_col, facet_page, id, pane, group, o, ...}{to be described}
}
\value{
internal tmap object
}
\description{
Internal methods for tm_donuts
}
\keyword{internal}
