% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridDonuts.R
\name{tmapGridDonuts}
\alias{tmapGridDonuts}
\alias{tmapLeafletDonuts}
\title{Internal methods for tm_donuts}
\usage{
tmapGridDonuts(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletDonuts(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{shpTM, dt, gp, bbx, facet_row, facet_col, facet_page, id, pane, group, o, ...}{to be described}
}
\description{
Internal methods for tm_donuts
}
\keyword{internal}
