% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridFlowers.R
\name{tmapGridFlowers}
\alias{tmapGridFlowers}
\alias{tmapLeafletFlowers}
\title{Internal methods for tm_flowers}
\usage{
tmapGridFlowers(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletFlowers(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{shpTM, dt, gp, bbx, facet_row, facet_col, facet_page, id, pane, group, o, ...}{to be described}
}
\value{
internal tmap object
}
\description{
Internal methods for tm_flowers
}
\keyword{internal}
