% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapScaleComposition.R
\name{tmapScaleComposition}
\alias{tmapScaleComposition}
\title{Internal method for tm_scale_composition}
\usage{
tmapScaleComposition(
  ...,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)
}
\arguments{
\item{...}{variables}

\item{scale, legend, chart, o, aes, layer, layer_args, sortRev, bypass_ord, submit_legend}{arguments to be described}
}
\description{
Internal method for tm_scale_composition
}
\keyword{internal}
