% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapScaleMulti.R, R/tmapValues_comppart.R
\name{tmapScaleMulti}
\alias{tmapScaleMulti}
\alias{tmapValuesCheck_multi}
\alias{tmapValuesIsDiv_multi}
\alias{tmapValuesRange_multi}
\alias{tmapValuesVV_multi}
\alias{tmapValuesSubmit_multi}
\alias{tmapValuesScale_multi}
\alias{tmapValuesColorize_multi}
\alias{tmapValuesCVV_multi}
\title{Internal method for tm_scale_multi}
\usage{
tmapScaleMulti(
  ...,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapValuesCheck_multi(x, is_var = TRUE)

tmapValuesIsDiv_multi(x)

tmapValuesRange_multi(x, n, isdiv)

tmapValuesVV_multi(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesSubmit_multi(x, args)

tmapValuesScale_multi(x, scale)

tmapValuesColorize_multi(x, pc)

tmapValuesCVV_multi(x, value.na, n, range, scale, rep, o)
}
\arguments{
\item{...}{variables}

\item{scale, legend, chart, o, aes, layer, layer_args, sortRev, bypass_ord, submit_legend}{arguments to be described}
}
\description{
Internal method for tm_scale_multi
}
\keyword{internal}
