% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapScaleComposition.R, R/tmapValues_comppart.R
\name{tmapScaleComposition}
\alias{tmapScaleComposition}
\alias{tmapValuesCheck_comppart}
\alias{tmapValuesIsDiv_comppart}
\alias{tmapValuesRange_comppart}
\alias{tmapValuesVV_comppart}
\alias{tmapValuesSubmit_comppart}
\alias{tmapValuesScale_comppart}
\alias{tmapValuesColorize_comppart}
\alias{tmapValuesCVV_comppart}
\title{Internal method for tm_scale_composition}
\usage{
tmapScaleComposition(
  ...,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapValuesCheck_comppart(x, is_var = TRUE)

tmapValuesIsDiv_comppart(x)

tmapValuesRange_comppart(x, n, isdiv)

tmapValuesVV_comppart(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesSubmit_comppart(x, args)

tmapValuesScale_comppart(x, scale)

tmapValuesColorize_comppart(x, pc)

tmapValuesCVV_comppart(x, value.na, n, range, scale, rep, o)
}
\arguments{
\item{...}{variables}

\item{scale, legend, chart, o, aes, layer, layer_args, sortRev, bypass_ord, submit_legend}{arguments to be described}
}
\value{
internal tmap object
}
\description{
Internal method for tm_scale_composition
}
\keyword{internal}
